Add sequential cornell box changes

+++ sequential/aabb.h
@@ -26,7 +26,7 @@ class aabb {
     
     aabb pad() {
         // Return an AABB that has no side narrower than some delta , padding if necessary.
-        double delta = 0.0001;
+        float delta = 0.0001;
         interval new_x = (x.size() >= delta) ? x : x.expand(delta);
         interval new_y = (y.size() >= delta) ? y : y.expand(delta);
         interval new_z = (z.size() >= delta) ? z : z.expand(delta);
